Household net income variables
