+++ policyengine_canada/parameters/gov/cra/tax/income/credits/non_refundable.yaml
@@ -1,0 +1,7 @@
+description: Non-refundable tax credits.
+values:
+  2019-01-01:
+    - age_amount_credit
+    - dtc
+metadata:
+  unit: program
