+++ policyengine_canada/variables/gov/cra/tax/income/credits/non_refundable_tax_credits.py
@@ -1,0 +1,11 @@
+from policyengine_canada.model_api import *
+
+
+class non_refundable_tax_credits(Variable):
+    value_type = float
+    entity = Household
+    label = "Non-refundable tax credits"
+    unit = CAD
+    definition_period = YEAR
+
+    formula = sum_of_variables("gov.cra.tax.income.credits.non_refundable")
