+++ policyengine_canada/variables/gov/cra/tax/income/income_tax_before_credits.py
@@ -1,10 +1,10 @@
 from policyengine_canada.model_api import *
 
 
-class income_tax(Variable):
+class income_tax_before_credits(Variable):
     value_type = float
     entity = Person
-    label = "Income tax"
+    label = "Income tax before non-refundable tax credits"
     unit = CAD
     documentation = "Example income tax regime"
     definition_period = YEAR
